requirements.txt
langchain-qdrant
langchain 
langchain-openai
qdrant-client
python-dotenv
pyowm
langchain-community
langchain-text-splitters
pypdf
langgraph
streamlit
langsmith
pytest